ref(api): Rename helper group index helper file
