+++ src/sentry/api/helpers/group_index.py
